完成：系统筛选推荐：系统根据客户提交的要求，快速筛选出符合条件的空闲会议室，并    展示相关信息，如会议室面积、座位数、设备详情、每小时租赁价格等，供用户选择。生成订单：客户选定会议室后，系统自动生成订单，会议室进入锁定状态。
   添加定时任务查出超过 30 分钟未支付的 pending_payment 状态订单并更新

+++ backend/src/main/java/com/BookRoom/entity/meetingRoom/MeetingRoom.java
@@ -5,6 +5,8 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @TableName("meetingRoom")
 public class MeetingRoom {
@@ -13,7 +15,7 @@ public class MeetingRoom {
     private String meetingRoomName;
     private String creatorId;
     private String departmentId;
-    private Double pricePerHour; // 原 credit 字段已删除，替换为 pricePerHour
+    private BigDecimal pricePerHour; // 原 credit 字段已删除，替换为 pricePerHour
     private String description;
     private String type;
     private Integer capacity;
